GH2883 moving class
