+++ common-service-core/src/main/java/org/zowe/apiml/services/ServiceInfo.java
@@ -8,7 +8,7 @@
  * Copyright Contributors to the Zowe Project.
  */
 
-package org.zowe.apiml.product.services;
+package org.zowe.apiml.services;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.netflix.appinfo.InstanceInfo;
